+++ src/common/Collision/Management/MMapManager.h
@@ -78,10 +78,6 @@ namespace MMAP
             uint32 getLoadedMapsCount() const { return uint32(loadedMMaps.size()); }
         private:
             bool loadMapData(std::string const& basePath, uint32 mapId);
-            bool loadMapImpl(std::string const& basePath, uint32 mapId, int32 x, int32 y);
-            bool loadMapInstanceImpl(std::string const& basePath, uint32 mapId, uint32 instanceId);
-            bool unloadMapImpl(uint32 mapId, int32 x, int32 y);
-            bool unloadMapImpl(uint32 mapId);
             uint32 packTileID(int32 x, int32 y);
 
             MMapDataSet::const_iterator GetMMapData(uint32 mapId) const;
@@ -89,7 +85,6 @@ namespace MMAP
             uint32 loadedTiles;
             bool thread_safe_environment;
 
-            std::unordered_map<uint32, std::vector<uint32>> childMapData;
             std::unordered_map<uint32, uint32> parentMapData;
     };
 }
