+++ src/common/Collision/Maps/MapTree.h
@@ -62,8 +62,9 @@ namespace VMAP
 
             struct TileFileOpenResult
             {
-                FILE* File;
                 std::string Name;
+                FILE* File;
+                int32 UsedMapId;
             };
 
         private:
@@ -85,9 +86,9 @@ namespace VMAP
             bool getAreaInfo(G3D::Vector3 &pos, uint32 &flags, int32 &adtId, int32 &rootId, int32 &groupId) const;
             bool GetLocationInfo(const G3D::Vector3 &pos, LocationInfo &info) const;
 
-            bool InitMap(std::string const& fname);
+            LoadResult InitMap(std::string const& fname);
             void UnloadMap(VMapManager2* vm);
-            bool LoadMapTile(uint32 tileX, uint32 tileY, VMapManager2* vm);
+            LoadResult LoadMapTile(uint32 tileX, uint32 tileY, VMapManager2* vm);
             void UnloadMapTile(uint32 tileX, uint32 tileY, VMapManager2* vm);
             uint32 numLoadedTiles() const { return uint32(iLoadedTiles.size()); }
             void getModelInstances(ModelInstance* &models, uint32 &count);
