+++ src/server/game/Handlers/QueryHandler.cpp
@@ -22,12 +22,12 @@
 #include "DBCStores.h"
 #include "GameTime.h"
 #include "Log.h"
-#include "MapManager.h"
 #include "NPCHandler.h"
 #include "ObjectAccessor.h"
 #include "ObjectMgr.h"
 #include "Player.h"
 #include "QueryPackets.h"
+#include "TerrainMgr.h"
 #include "UpdateMask.h"
 #include "World.h"
 #include "WorldPacket.h"
@@ -150,12 +150,12 @@ void WorldSession::HandleCorpseQueryOpcode(WorldPacket& /*recvData*/)
             if (corpseMapEntry->IsDungeon() && corpseMapEntry->CorpseMapID >= 0)
             {
                 // if corpse map have entrance
-                if (Map* entranceMap = sMapMgr->CreateBaseMap(corpseMapEntry->CorpseMapID))
+                if (std::shared_ptr<TerrainInfo> entranceTerrain = sTerrainMgr.LoadTerrain(corpseMapEntry->CorpseMapID))
                 {
                     mapID = corpseMapEntry->CorpseMapID;
                     x = corpseMapEntry->Corpse.X;
                     y = corpseMapEntry->Corpse.Y;
-                    z = entranceMap->GetHeight(GetPlayer()->GetPhaseShift(), x, y, MAX_HEIGHT);
+                    z = entranceTerrain->GetStaticHeight(GetPlayer()->GetPhaseShift(), x, y, MAX_HEIGHT);
                 }
             }
         }
