+++ src/server/game/Scripting/ScriptMgr.h
@@ -38,7 +38,6 @@ class DynamicObject;
 class GameObject;
 class GameObjectAI;
 class Guild;
-class GridMap;
 class Group;
 class InstanceMap;
 class InstanceScript;
@@ -322,12 +321,6 @@ template<class TMap> class MapScript : public UpdatableScript<TMap>
 
         // Called just before the map is destroyed.
         virtual void OnDestroy(TMap* /*map*/) { }
-
-        // Called when a grid map is loaded.
-        virtual void OnLoadGridMap(TMap* /*map*/, GridMap* /*gmap*/, uint32 /*gx*/, uint32 /*gy*/) { }
-
-        // Called when a grid map is unloaded.
-        virtual void OnUnloadGridMap(TMap* /*map*/, GridMap* /*gmap*/, uint32 /*gx*/, uint32 /*gy*/)  { }
 
         // Called when a player enters the map.
         virtual void OnPlayerEnter(TMap* /*map*/, Player* /*player*/) { }
@@ -898,8 +891,6 @@ class TC_GAME_API ScriptMgr
 
         void OnCreateMap(Map* map);
         void OnDestroyMap(Map* map);
-        void OnLoadGridMap(Map* map, GridMap* gmap, uint32 gx, uint32 gy);
-        void OnUnloadGridMap(Map* map, GridMap* gmap, uint32 gx, uint32 gy);
         void OnPlayerEnterMap(Map* map, Player* player);
         void OnPlayerLeaveMap(Map* map, Player* player);
         void OnMapUpdate(Map* map, uint32 diff);
