+++ src/server/worldserver/Main.cpp
@@ -47,6 +47,7 @@
 #include "ScriptMgr.h"
 #include "ScriptReloadMgr.h"
 #include "TCSoap.h"
+#include "TerrainMgr.h"
 #include "World.h"
 #include "WorldSocket.h"
 #include "WorldSocketMgr.h"
@@ -301,6 +302,7 @@ extern int main(int argc, char** argv)
         sInstanceSaveMgr->Unload();
         sOutdoorPvPMgr->Die();                     // unload it before MapManager
         sMapMgr->UnloadAll();                      // unload all grids (including locked in memory)
+        sTerrainMgr.UnloadAll();
     });
 
     // Start the Remote Access port (acceptor) if enabled
