+++ src/tools/mmaps_generator/TerrainBuilder.cpp
@@ -588,12 +588,12 @@ namespace MMAP
     bool TerrainBuilder::loadVMap(uint32 mapID, uint32 tileX, uint32 tileY, MeshData &meshData)
     {
         std::unique_ptr<VMapManager2> vmapManager = VMapFactory::CreateVMapManager();
-        int result = vmapManager->loadSingleMap(mapID, "vmaps", tileX, tileY);
+        LoadResult result = vmapManager->loadMap("vmaps", mapID, tileX, tileY);
         bool retval = false;
 
         do
         {
-            if (result == VMAP_LOAD_RESULT_ERROR)
+            if (result != LoadResult::Success)
                 break;
 
             InstanceTreeMap instanceTrees;
@@ -732,7 +732,7 @@ namespace MMAP
         }
         while (false);
 
-        vmapManager->unloadSingleMap(mapID, tileX, tileY);
+        vmapManager->unloadMap(mapID, tileX, tileY);
 
         return retval;
     }
